Fix problems referencing partial 'global/nav' from layout
Layout didn't get the app id appended to it, so referencing a partial
with a name 'global/nav' resulted in the monkeypatching expecting that
the app id was 'global'. Additionally, even if the monkeypatched vesion
received a valid raw path, the raw version was never attempted.

This is fixed in two ways:

  * First, we add the app id to the layout sent to any 'render' call
    except for `layout: false`
  * Second, we add a fallback to the monkeypatching where we simply try
    what express would do with the raw path parameter

The tests show both problems.

Of note is the fact that the problem can be fixed by #2 alone, but then
it always checks the filesystem twice. The change for #1 avoids that.